requirements.txt
torch
datasets
elasticsearch==8.12.1
tqdm
python-dotenv

pandas
fastapi
uvicorn
langchain
langchain-google-genai
google-generativeai
langchain-core
langchain-text-splitters
# Your RAG libraries:
sentence-transformers
faiss-cpu
transformers
pypdf
python-docx
python-pptx
# Tool libraries:
fpdf2
reportlab